+++ changelog.md
@@ -3,6 +3,7 @@
 ## Upcoming
 * Further reduced Ender Chests hardness
 * Reduced witches potions drop (~~2\~4~~ -> 0\~4 potions per witch)
+* Knockback multipliers are now configurable through json
 
 ## 3.9.10
 * Reduced hoes base damage (~~2~~ -> 1.5)